MessageChoices
- Moved messagechoices to a utils
	- Allows flexibility with testing/validation
